added enemy chase and attack functionality
now after the correct ability is used on a fighting trap the enemy will start chasing the player and attacking on some time interval. if the player runs away far enough the enemy will go back to it's location.
if the player attacks the enemy with some ability the attack interval will reset.
on stun enemy stops in place, on poison its speed is lower by 0.75.
made new class AttackCast from the class that was inside the player script.